+++ asdf/tests/helpers.py
@@ -392,10 +392,16 @@ def assert_no_warnings(warning_class=None):
     import pytest
 
     if warning_class is None:
-        with warnings.catch_warnings():
-            warnings.simplefilter("error")
+        # check if warnings already produce errors
+        for f in warnings.filters:
+            if f[0] == "error":
+                yield
+                break
+        else:
+            with warnings.catch_warnings():
+                warnings.simplefilter("error")
 
-            yield
+                yield
     else:
         with pytest.warns(Warning) as recorded_warnings:
             yield