+++ asdf/tests/test_deprecated.py
@@ -7,9 +7,11 @@ import asdf._types
 import asdf.extension
 import asdf.testing.helpers
 from asdf._types import CustomType
+from asdf import entry_points
 from asdf.exceptions import AsdfDeprecationWarning
 from asdf.tests.helpers import assert_extension_correctness
 from asdf.tests.objects import CustomExtension
+from .test_entry_points import _monkeypatch_entry_points, mock_entry_points  # noqa: F401
 
 
 def test_custom_type_warning():
@@ -103,3 +105,9 @@ def test_extension_class_deprecation(name):
 def test_top_level_asdf_extension_deprecation():
     with pytest.warns(AsdfDeprecationWarning, match="AsdfExtension is deprecated"):
         asdf.AsdfExtension
+
+
+def test_deprecated_entry_point(mock_entry_points):  # noqa: F811
+    mock_entry_points.append(("asdf_extensions", "legacy", "asdf.tests.test_entry_points:LegacyExtension"))
+    with pytest.warns(AsdfDeprecationWarning, match="^.*LegacyExtension.* uses the deprecated.*$"):
+        entry_points.get_extensions()